requirements-dev.txt
-r requirements.txt
black==19.10b0
pylint==2.6.0
pylint-django==2.3.0
pytest-django==3.10.0
pytest-cov==2.10.1
pytest-xdist==2.1.0
jinjalint==0.5
attrs==20.2.0
django-debug-toolbar==3.1.1